edit: move components to src/components, edited Podiums.tsx and ScoreDisplay.tsx, delete duplicate font
