+++ src/components/Podiums.tsx
@@ -31,7 +31,7 @@ const colorMap: { [key: string]: { [key: string]: string } } = {
 };
 
 // Podium item
-const Podium = (props: {
+const Podium = ({ color, name, points, place }: {
   color: string;
   name: string;
   points: number;
@@ -40,32 +40,32 @@ const Podium = (props: {
   return (
     <div
       className="w-[100px] h-[350px] text-xl rounded-lg justify-start flex flex-col items-center"
-      style={{ backgroundColor: colorMap[props.color].bg }}
+      style={{ backgroundColor: colorMap[color].bg }}
     >
       <div className="w-full h-auto flex items-center justify-center">
-        <Star color={colorMap[props.color].star} />
+        <Star color={colorMap[color].star} />
         <p
           className="absolute  text-4xl font-bold font-roboto-condensed mt-3 -ml-4"
-          style={{ color: colorMap[props.color].textStar }}
+          style={{ color: colorMap[color].textStar }}
         >
-          {props.place[0]}
+          {place[0]}
         </p>
         <p
           className="absolute text-lg font-bold font-roboto-condensed -mt-1 ml-5"
-          style={{ color: colorMap[props.color].textStar }}
+          style={{ color: colorMap[color].textStar }}
         >
-          {props.place[1]}
+          {place[1]}
         </p>
       </div>
 
-      <p>{props.name}</p>
-      <p className="font-roboto-condensed font-semibold">{props.points} pt</p>
+      <p>{name}</p>
+      <p className="6font-roboto-condensed">{points} pts</p>
     </div>
   );
 };
 
 // star svg
-const Star = (props: { color: string }) => {
+const Star = ({color}: { color: string }) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -77,7 +77,7 @@ const Star = (props: { color: string }) => {
     >
       <path
         d="M39 0L53.8122 19.6128L77.0423 27.6393L62.9666 47.7872L62.5114 72.3607L39 65.2L15.4886 72.3607L15.0334 47.7872L0.957741 27.6393L24.1878 19.6128L39 0Z"
-        fill={props.color}
+        fill={color}
       />
     </svg>
   );
