+++ src/components/ScoreDisplay.tsx
@@ -4,21 +4,21 @@ interface mockDataInterface {
   score: number;
 }
 
-export default function ScoreDisplay() {
+export default function ScoreDisplay({ start }:{ start: number }) {
   return (
-    <div className="w-[90%] relative flex flex-col font-athiti items-center font-[450] text-xl text-blue-1 ">
+    <div className="w-[90%] relative flex flex-col font-athiti items-center font-[450] text-xl text-blue-1 mx-auto">
       <table className="w-[90%] max-w-[350px]">
         <tbody>
-          {mockData.map((item, index: number) => (
+          {mockData.slice(start-1).map((item, index: number) => (
             <tr key={index} className="w-full">
               <td className="w-[20%]">
                 <div className="w-full relative flex flex-row space-x-1/2">
-                  <p>{index + 3}</p>
+                  <p>{index + start}</p>
                   <p className="relative top-0 text-sm ">th</p>
                 </div>
               </td>
-              <td className="w-[60%]">{item.name}</td>
-              <td className="w-[40%]">{item.score}</td>
+              <td className="w-[50%]">{item.name}</td>
+              <td className="w-[50%]">{item.score} pts</td>
             </tr>
           ))}
         </tbody>
